requirements.txt
opencv-python
pytesseract
Pillow
numpy
python-dotenv
pytest
